Added a separate page for backup options
