+++ mobile/lib/routing/router.gr.dart
@@ -373,6 +373,12 @@ class _$AppRouter extends RootStackRouter {
         barrierDismissible: false,
       );
     },
+    BackupOptionsRoute.name: (routeData) {
+      return MaterialPageX<dynamic>(
+        routeData: routeData,
+        child: const BackupOptionsPage(),
+      );
+    },
     HomeRoute.name: (routeData) {
       return MaterialPageX<dynamic>(
         routeData: routeData,
@@ -720,6 +726,14 @@ class _$AppRouter extends RootStackRouter {
         RouteConfig(
           MapLocationPickerRoute.name,
           path: '/map-location-picker-page',
+          guards: [
+            authGuard,
+            duplicateGuard,
+          ],
+        ),
+        RouteConfig(
+          BackupOptionsRoute.name,
+          path: '/backup-options-page',
           guards: [
             authGuard,
             duplicateGuard,
@@ -1674,6 +1688,18 @@ class MapLocationPickerRouteArgs {
   String toString() {
     return 'MapLocationPickerRouteArgs{key: $key, initialLatLng: $initialLatLng}';
   }
+}
+
+/// generated route for
+/// [BackupOptionsPage]
+class BackupOptionsRoute extends PageRouteInfo<void> {
+  const BackupOptionsRoute()
+      : super(
+          BackupOptionsRoute.name,
+          path: '/backup-options-page',
+        );
+
+  static const String name = 'BackupOptionsRoute';
 }
 
 /// generated route for
